Agregue mas validaciones, PagoPrestamo y SolicitarPrestamo
Ahora es "funcional" en teoria, lo voy a probar cuando se implemente la
web.
Falta la validacion de cantidad de cuotas que se necesita soporte de
base de datos.

Agregado enum de PAGO_PRESTAMO, para saber si es efectivo o cuenta.

+++ RamonBank/src/com/ramon/ramonbank/dbaccess/tables/PagoPrestamos.java
@@ -22,7 +22,7 @@ public class PagoPrestamos extends Tables {
 	private int _origen;
 	private int _idPrestamo;
 	private int _cantCuotas;
-	private int _monto;
+	private double _monto;
 	private Fecha _filtro_fechaDesde;
 	private Fecha _filtro_fechaHasta;
 	
@@ -96,11 +96,11 @@ public class PagoPrestamos extends Tables {
 		this._cantCuotas = _cantCuotas;
 	}
 
-	public int get_monto() {
+	public double get_monto() {
 		return _monto;
 	}
 
-	public void set_monto(int _monto) {
+	public void set_monto(double _monto) {
 		this._monto = _monto;
 	}
